api/requirements.txt
# Django and DRF
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-filter>=23.0
drf-spectacular>=0.26.0

# Database
psycopg2-binary==2.9.9

# Web server
gunicorn>=21.0.0
whitenoise>=6.5.0

# Data processing
pandas==2.1.4
numpy==1.26.4

# Configuration
python-dotenv==1.0.0

# Utilities
python-dateutil>=2.8.2
tqdm==4.67.1

# Development (optionnel)
django-debug-toolbar>=4.0.0
pytest-django>=4.5.0 